+++ backend/models/publishingCompany.js
@@ -1,8 +1,0 @@
-const mongoose = require("mongoose");
-
-const publishingCompanySchema = new mongoose.Schema({
-  _id: mongoose.Schema.Types.ObjectId,
-  name: String,
-});
-
-module.exports = mongoose.model("PublishingCompany", publishingCompanySchema);